change price type to decimal

+++ BookHub/BusinessLayer/Services/IOrderService.cs
@@ -6,7 +6,7 @@ namespace BusinessLayer.Services;
 public interface IOrderService
 {
     Task<IEnumerable<OrderDetail>> GetOrdersAsync(int? userId, string? username,
-        DateTime? startDate, DateTime? endDate, double? totalPrice, int? bookId, string? bookName);
+        DateTime? startDate, DateTime? endDate, decimal? totalPrice, int? bookId, string? bookName);
     Task<OrderDetail> GetOrderByIdAsync(int id);
     Task<OrderDetail> CreateOrderAsync(OrderCreate orderCreate);
     Task<OrderDetail> UpdateOrderAsync(int id, OrderUpdate orderUpdate);
